fix polygon2

+++ include/math/polygon2.h
@@ -6,8 +6,8 @@
 
 namespace mt
 {
-    void polygon2_centroid( const mt::vec2f * _polygon, uint32_t _count, mt::vec2f * const _point );
-    bool polygon2_intersect( const mt::vec2f * _polygon, uint32_t _count, const mt::vec2f & _point );
-    bool polygon2_intersect( const mt::vec2f * _polygon, uint32_t _count, const mt::vec2f & _point, float _rx, float _ry );
-    bool polygon2_intersect( const mt::vec2f * _polygon1, uint32_t _count1, const mt::vec2f * _polygon2, uint32_t _count2 );
+    void polygon2_centroid( const mt::vec2f * _polygon, size_t _count, mt::vec2f * const _point );
+    bool polygon2_intersect( const mt::vec2f * _polygon, size_t _count, const mt::vec2f & _point );
+    bool polygon2_intersect( const mt::vec2f * _polygon, size_t _count, const mt::vec2f & _point, float _rx, float _ry );
+    bool polygon2_intersect( const mt::vec2f * _polygon1, size_t _count1, const mt::vec2f * _polygon2, size_t _count2 );
 }
